Bulk of design overhaul
